requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.2.1
cs50==9.4.0
Flask==3.1.1
Flask-Session==0.8.0
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
packaging==25.0
python-dotenv==1.1.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
termcolor==3.1.0
typing_extensions==4.14.1
Werkzeug==3.1.3
wheel==0.45.1
